requirements.txt
# Required packages
fastapi
pydantic
uvicorn
pandas
scikit-learn
openpyxl
pyspellchecker
nltk
joblib